Update README.md
This project aims to develop a sophisticated smart fire alarm system that addresses these limitations by leveraging advanced microcontroller technology, wireless communication, and GPS capabilities.The advent of Internet of Things (IoT) technology has revolutionized the approach to fire safety. IoT-enabled fire alarm systems provide accurate fire point localization, real-time monitoring, fire tracking, online updates, and simplified maintenance. Various research efforts have focused on integrating wireless communication technologies into indoor fire protection systems. 

The core of this smart fire alarm system is built around the Tiva C microcontroller, which is renowned for its performance and versatility in embedded systems. The Tiva C microcontroller serves as the primary processing unit, interfacing with various components including a fire detection sensor, a buzzer for local alerts, an ESP32 module for wireless communication, and a GPS module for location tracking.

The system is designed to detect fire through a connected sensor, which provides real-time data to the Tiva C microcontroller. To mitigate the risk of false alarms, a five-second delay is introduced upon detection of potential fire conditions. This delay allows the system to verify the presence of fire before triggering the alarm. Once the presence of fire is confirmed, the system activates a buzzer to provide immediate auditory alerts on-site, ensuring that occupants are promptly warned.
In addition to local alerts, the smart fire alarm system is equipped with remote notification capabilities. The ESP32 module, a powerful Wi-Fi enabled microcontroller, is utilized to send email notifications to designated recipients. This ensures that even when the occupants are not present, they are promptly informed of the potential fire hazard. The email notification includes crucial details such as the longitude and latitude of the current location, obtained from the integrated GPS module. This precise location information aids in directing emergency services to the exact site of the incident swiftly.

Key features of this fire alarm system include reliable communication between components, mechanisms to reduce false alarms, and considerations for power stability. The use of the GPS module enhances the system's capability by providing accurate location data, which is crucial for emergency responders. The ESP32's Wi-Fi connectivity facilitates seamless email notifications, ensuring that alerts are received promptly, even if the occupants are not present.

Overall, this project exemplifies the fusion of modern microcontroller technology, wireless communication, and location tracking to create a comprehensive fire safety solution. The smart fire alarm system is poised to significantly improve response times and protection measures in both residential and commercial settings, marking a substantial advancement in fire safety technology